[brain tests] Burst typing extensions from the main file
