+++ tests/brain/test_brain.py
@@ -23,15 +23,6 @@ from astroid.exceptions import (
 from astroid.nodes.node_classes import Const
 from astroid.nodes.scoped_nodes import ClassDef
 
-try:
-    import typing_extensions  # pylint: disable=unused-import
-
-    HAS_TYPING_EXTENSIONS = True
-    HAS_TYPING_EXTENSIONS_TYPEVAR = hasattr(typing_extensions, "TypeVar")
-except ImportError:
-    HAS_TYPING_EXTENSIONS = False
-    HAS_TYPING_EXTENSIONS_TYPEVAR = False
-
 
 def assertEqualMro(klass: ClassDef, expected_mro: list[str]) -> None:
     """Check mro names."""
@@ -1517,29 +1508,6 @@ class TypingBrain(unittest.TestCase):
         i1 = next(ast_nodes[1].infer())
         assert isinstance(i1, nodes.Const)
         assert i1.value == 2  # should be "Hello"!
-
-
-@pytest.mark.skipif(
-    not HAS_TYPING_EXTENSIONS,
-    reason="These tests require the typing_extensions library",
-)
-class TestTypingExtensions:
-    @staticmethod
-    @pytest.mark.skipif(
-        not HAS_TYPING_EXTENSIONS_TYPEVAR,
-        reason="Need typing_extensions>=4.4.0 to test TypeVar",
-    )
-    def test_typing_extensions_types() -> None:
-        ast_nodes = builder.extract_node(
-            """
-        from typing_extensions import TypeVar
-        TypeVar('MyTypeVar', int, float, complex) #@
-        TypeVar('AnyStr', str, bytes) #@
-        """
-        )
-        for node in ast_nodes:
-            inferred = next(node.infer())
-            assert isinstance(inferred, nodes.ClassDef)
 
 
 class ReBrainTest(unittest.TestCase):
